requirements.txt
altair==4.2.0
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
brotlipy==0.7.0
cachetools==5.1.0
certifi==2022.5.18.1
# cffi @ file:///home/builder/ci_310/cffi_1642753365720/work
# charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
ci-sdr==0.0.0
click==8.0.4
ConfigArgParse==1.5.3
# cryptography @ file:///tmp/build/80754af9/cryptography_1652101588599/work
ctc-segmentation==1.7.1
Cython==0.29.30
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
Distance==0.1.3
einops==0.4.1
entrypoints==0.4
espnet==202204
espnet-model-zoo==0.1.7
espnet-tts-frontend==0.0.3
executing==0.8.3
fastapi==0.78.0
fastjsonschema==2.15.3
filelock==3.7.0
g2p-en==2.1.0
gitdb==4.0.9
GitPython==3.1.27
h11==0.13.0
h5py==3.7.0
huggingface-hub==0.7.0
humanfriendly==10.0
# idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.19.2
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.4
inflect==5.6.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jaconv==0.3
jamo==0.4.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kaldiio==2.17.2
librosa==0.9.1
llvmlite==0.38.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
# mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
MouseInfo==0.1.3
moviepy==1.0.3
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.11
numba==0.55.1
numpy==1.21.6
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pooch==1.6.0
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
# pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.9.1
pydeck==0.7.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.12.0
Pympler==1.0.1
PyMsgBox==1.0.9
# pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyperclip==1.8.2
pypinyin==0.44.0
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
# PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python3-xlib==0.15
pytorch-wpe==0.0.1
pytube==12.1.0
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyworld==0.2.12
PyYAML==6.0
pyzmq==23.0.0
regex==2022.4.24
# requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
resampy==0.2.2
scikit-learn==1.1.1
scipy==1.8.1
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.96
# six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.2.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
sox==1.4.1
stack-data==0.2.0
starlette==0.19.1
streamlit==1.9.0
streamlit-player==0.1.5
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
toolz==0.11.2
torch==1.11.0
torch-complex==0.4.3
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
transformers==4.19.2
typeguard==2.13.3
# typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
tzdata==2022.1
tzlocal==4.2
Unidecode==1.3.4
# urllib3 @ file:///opt/conda/conda-bld/urllib3_1650637206367/work
uvicorn==0.17.6
validators==0.19.0
watchdog==2.1.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
# youtube-dl @ file:///home/conda/feedstock_root/build_artifacts/youtube-dl_1640026230460/work
zipp==3.8.0
